requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
requests
beautifulsoup4
google.generativeai